reading block info & dd sp sizelit strat

+++ pixels-core/src/test/java/cn/edu/ruc/iir/pixels/core/writer/TestPixelsWriter.java
@@ -54,7 +54,7 @@ public class TestPixelsWriter
             PixelsWriter pixelsWriter =
                     PixelsWriterImpl.newBuilder()
                             .setSchema(schema)
-                            .setPixelStride(TestParams.pixelStride)
+                            .setPixelStride((int) TestParams.pixelStride)
                             .setRowGroupSize(TestParams.rowGroupSize)
                             .setFS(fs)
                             .setFilePath(new Path(filePath))
@@ -62,7 +62,7 @@ public class TestPixelsWriter
                             .setReplication(TestParams.blockReplication)
                             .setBlockPadding(TestParams.blockPadding)
                             .setEncoding(TestParams.encoding)
-                            .setCompressionBlockSize(TestParams.compressionBlockSize)
+                            .setCompressionBlockSize((int) TestParams.compressionBlockSize)
                             .build();
 
             long curT = System.currentTimeMillis();
